refactor: removed unused index.ts files & fix imports & shared/ui refactor
