+++ src/shared/ui/card/ui/Card.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../lib/utils/cn';
+import { cn } from '../../../lib/utils/cn';
 
 import type { CardProps } from './Card.types';
