+++ src/shared/ui/card/ui/Card.types.ts
