--- src/shared/ui/index.ts
@@ -1,20 +1,0 @@
-export { BackButton } from './back-button';
-export { Button, buttonVariants } from './button';
-export {
-  Card,
-  CardAction,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from './card';
-export { Container } from './container';
-export { ErrorField } from './error-field';
-export { FormField } from './form-field';
-export { Input } from './input';
-export { Label } from './label';
-export { Select } from './select';
-export { Separator } from './separator';
-export { Skeleton } from './skeleton';
-export { Textarea } from './text-area';
