+++ src/shared/ui/input/index.ts
@@ -1,2 +1,2 @@
-export { Input } from './Input';
-export type { InputProps } from './Input.types';
+export { Input } from './ui/Input';
+export type { InputProps } from './ui/Input.types';
