+++ src/shared/ui/label/ui/Label.types.ts
