+++ src/shared/ui/separator/index.ts
@@ -1,2 +1,2 @@
-export { Separator } from './Separator';
-export type { SeparatorProps } from './Separator.types';
+export { Separator } from './ui/Separator';
+export type { SeparatorProps } from './ui/Separator.types';
