+++ src/shared/ui/separator/ui/Separator.types.ts
