+++ src/shared/ui/skeleton/ui/Skeleton.types.ts
