+++ src/shared/ui/text-area/ui/Textarea.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../lib/utils/cn';
+import { cn } from '../../../lib/utils/cn';
 
 import type { TextareaProps } from './Textarea.types';
